# swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/modulos": {
            "post": {
                "operationId": "ModuloController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModuloDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "modulos"
                ]
            },
            "get": {
                "operationId": "ModuloController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": " query-Filter.e.g.col1:v1,col2:v2… ",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "modulos"
                ]
            }
        },
        "/modulos/{id}": {
            "get": {
                "operationId": "ModuloController_getById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "modulos"
                ]
            },
            "put": {
                "operationId": "ModuloController_put",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModuloDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "modulos"
                ]
            },
            "delete": {
                "operationId": "ModuloController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "modulos"
                ]
            }
        },
...
        "/modal-alertas": {
            "post": {
                "operationId": "ModalAlertaController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModalAlertaDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "modal-alertas"
                ]
            },
            "get": {
                "operationId": "ModalAlertaController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": " query-Filter.e.g.col1:v1,col2:v2… ",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "modal-alertas"
                ]
            }
        },
        "/modal-alertas/{id}": {
            "get": {
                "operationId": "ModalAlertaController_getById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "modal-alertas"
                ]
            },
            "put": {
                "operationId": "ModalAlertaController_put",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModalAlertaDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "modal-alertas"
                ]
            },
            "delete": {
                "operationId": "ModalAlertaController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "modal-alertas"
                ]
            }
        },
...
    },
    "info": {
        "title": "plantillas_formularios_dinamicos_crud",
        "description": "API CRUD para la gestion de plantillas para la creacion de formularios dinamicos",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
            "ModuloDto": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "sistema_id": {
                        "type": "number"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "fecha_creacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_modificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "nombre",
                    "descripcion",
                    "sistema_id",
                    "activo",
                    "fecha_creacion",
                    "fecha_modificacion"
                ]
            },
...
            "ModalAlertaDto": {
                "type": "object",
                "properties": {
                    "titulo": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "formulario_id": {
                        "type": "string"
                    },
                    "titulo_boton_principal": {
                        "type": "string"
                    },
                    "titulo_boton_secundario": {
                        "type": "string"
                    },
                    "tipo_id": {
                        "type": "number"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "fecha_creacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_modificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "titulo",
                    "descripcion",
                    "formulario_id",
                    "titulo_boton_principal",
                    "titulo_boton_secundario",
                    "tipo_id",
                    "activo",
                    "fecha_creacion",
                    "fecha_modificacion"
                ]
            },
...
        }
    }
}